backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
pypdf==4.3.1

# LangChain core
langchain==0.2.16
langchain-community==0.2.16

# Google / Gemini
langchain-google-genai==1.0.7
google-generativeai==0.7.2

# Vector store for RAG
faiss-cpu==1.8.0